+++ mariadb-c/mariadb_c/mariadb_connection.c
@@ -1140,7 +1140,8 @@ static PyObject *MrdbConnection_escape_string(MrdbConnection *self,
         PyObject *str)
 {
     PyObject *new_string= NULL;
-    size_t from_length, to_length;
+    Py_ssize_t from_length= 0;
+    size_t to_length;
     char *from, *to;
 
     /* escaping depends on the server status, so we need a valid
@@ -1152,8 +1153,11 @@ static PyObject *MrdbConnection_escape_string(MrdbConnection *self,
         return NULL;
     }
 
-    from= (char *)PyUnicode_AsUTF8AndSize(str, (Py_ssize_t *)&from_length);
-    if (!(to= (char *)PyMem_Calloc(1, from_length * 2 + 1)))
+    from= (char *)PyUnicode_AsUTF8AndSize(str, &from_length);
+    if (!from)
+        return NULL;
+
+    if (!(to= (char *)PyMem_Calloc(1, (size_t)from_length * 2 + 1)))
     {
         return NULL;
     }